Migrate to `dtest` for testing
