+++ modules/relative-path.js
@@ -3,7 +3,7 @@ export default function relative (from, to) {
   var toParts = to ? to.split('/') : []
   var relParts = []
   var baseCount = 0
-  
+
   for (var i = 0; i < toParts.length; i++) {
     if (fromParts[i] !== toParts[i]) {
       break
@@ -13,20 +13,6 @@ export default function relative (from, to) {
   for (var i = 0; i < fromParts.length - baseCount; i++) {
     relParts.push('..')
   }
-  
-  return relParts.concat(toParts.slice(baseCount)).join('/')
-}
 
-/**
- * Tests
- */
-export function test (t) {
-  var cases = [
-    "relative('arf/barf/gnarf', 'arf/yarf/blarf') === '../../yarf/blarf'",
-    "relative('barf/gnarf', 'arf/yarf') === '../../arf/yarf'",
-    "relative('barf/gnarf/yarf', 'barf/gnarf') === '..'",
-    "relative('', 'ping/pong') === 'ping/pong'"
-  ]
-  
-  cases.forEach(c => t.ok(eval(c), c))
+  return relParts.concat(toParts.slice(baseCount)).join('/')
 }
